+++ src/200Responses.ts
@@ -1,27 +1,27 @@
 import { APIGatewayProxyEvent } from 'aws-lambda';
-import { IObject, Response } from './BaseResponse';
-import { StatusCodes } from './utils';
+import { Response } from './BaseResponse';
+import { IOptions, StatusCodes } from './utils';
 
 export class Ok extends Response {
-    constructor(event: APIGatewayProxyEvent, body?: any, headers?: IObject, hideBody?: boolean) {
-        super(event, StatusCodes.OK, body, headers, hideBody);
+    constructor(event: APIGatewayProxyEvent, body?: any, options?: IOptions) {
+        super(event, StatusCodes.OK, body, options);
     }
 }
 
 export class Created extends Response {
-    constructor(event: APIGatewayProxyEvent, id?: any, headers?: IObject, hideBody?: boolean) {
-        super(event, StatusCodes.CREATED, id, headers, hideBody);
+    constructor(event: APIGatewayProxyEvent, id?: any, options?: IOptions) {
+        super(event, StatusCodes.CREATED, id, options);
     }
 }
 
 export class Accepted extends Response {
-    constructor(event: APIGatewayProxyEvent, body?: any, headers?: IObject, hideBody?: boolean) {
-        super(event, StatusCodes.ACCEPTED, body, headers, hideBody);
+    constructor(event: APIGatewayProxyEvent, body?: any, options?: IOptions) {
+        super(event, StatusCodes.ACCEPTED, body, options);
     }
 }
 
 export class NoContent extends Response {
-    constructor(event: APIGatewayProxyEvent, headers?: IObject, hideBody?: boolean) {
-        super(event, StatusCodes.NO_CONTENT, undefined, headers, hideBody);
+    constructor(event: APIGatewayProxyEvent, options?: IOptions) {
+        super(event, StatusCodes.NO_CONTENT, undefined, options);
     }
 }