+++ src/BaseResponse.ts
@@ -1,53 +1,35 @@
 import { APIGatewayProxyEvent } from 'aws-lambda';
 import * as jwt from 'jsonwebtoken';
-import { IOptions, StatusCodeType } from './utils';
+import { IOptions, IOptionsDefined, Options, StatusCodeType } from './utils';
 
-export interface IObject {
-    [key: string]: any;
-}
 
 export interface IResponse {
     statusCode: number;
     body: any;
-    headers: IObject
+    headers: {[key: string]: any;};
 }
 
 export abstract class Response implements IResponse {
     readonly statusCode: number;
     readonly body: any;
-    readonly headers: IObject = {
-        'Access-Control-Allow-Origin': '*',
-        'Access-Control-Allow-Credentials': true
-    };
+    readonly headers: {[key: string]: any};
+    private readonly options: IOptionsDefined;
 
-    constructor(
-        event: APIGatewayProxyEvent,
-        statusCode: StatusCodeType,
-        body?: any,
-        options: IOptions = {
-            hideBody: false,
-            logger: console,
-            tokenHeaderKey: 'Authorization'
-        }) {
+    constructor(event: APIGatewayProxyEvent, statusCode: StatusCodeType, body?: any, options?: IOptions) {
         this.statusCode = statusCode;
         this.body = JSON.stringify(body);
-        if (options.headers) {
-            this.headers = options.headers;
-        }
+        this.options = new Options(options);
+        this.headers = this.options.headers;
 
-        this.logRequest(event, options);
+        this.logRequest(event);
     }
 
-    private logRequest(event: APIGatewayProxyEvent, options: IOptions) {
-        let user;
+    private logRequest(event: APIGatewayProxyEvent) {
+        const token = event.headers[this.options.tokenHeaderKey];
 
-        if (event.headers[options.tokenHeaderKey!]) {
-            user = this.getUser(event.headers[options.tokenHeaderKey!]!);
-        }
-
-        options.logger.log(
+        this.options.logger.log(
             JSON.stringify({
-                user,
+                user: token ? this.getUser(token) : undefined,
                 responseStatus: this.statusCode,
                 timestamp: new Date(event.requestContext.requestTimeEpoch),
                 method: event.httpMethod,
@@ -55,21 +37,19 @@ export abstract class Response implements IResponse {
                 pathParameters: event.pathParameters,
                 query: event.queryStringParameters,
                 sourceIP: event.requestContext.identity.sourceIp,
-                body: options.hideBody ? '********' : event.body
+                body: this.options.hideBody ? '********' : event.body
             })
         );
     }
 
-    private getUser(token: string) {
-        const tokenUserKey = process.env.ALR_TOKEN_USER_KEY ?? 'email';
+    private getUser(token: string): string {
         const decodedToken = jwt.decode(token);
-
-        const user = (decodedToken as jwt.JwtPayload)[tokenUserKey];
+        const user = (decodedToken as jwt.JwtPayload)[this.options.tokenUserKey];
 
         if (user) {
             return user;
         } else {
-            throw new Error(`JWT invalid. Does not contain property ${tokenUserKey}`);
+            throw new Error(`JWT invalid. Does not contain property ${this.options.tokenUserKey}`);
         }
     }
 }